+++ docs/CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [[0.15.2]] - 2022-01-15 _(MJ)_
+
+###### _([diff: 0.15.1-0.15.2])_
+
+Added:
+
+- `UserFlowHandler` class
+
+Changed:
+
+- Moved globally scoped HTMLElements to `UserFlowHandler` class
+- Moved globally scoped Fullscreen related functions to `UserFlowHandler` -> `Fullscreen` class
+
 ## [[0.15.1]] - 2022-01-15 _(MJ)_
 
 ###### _([diff: 0.15.0-0.15.1])_
@@ -389,6 +402,7 @@ Added:
 - Created [**`CHANGELOG.md`**](../docs/CHANGELOG.md) for the project that contains the history of changes for this project.
 - Created [**`README.md`**](../README.md) for the project that contains various useful information, requirements and instructions in order ot run the program.
 
+[diff: 0.15.1-0.15.2]: https://github.com/Luzkan/CryptoImage/compare/0.15.1...0.15.2
 [diff: 0.15.0-0.15.1]: https://github.com/Luzkan/CryptoImage/compare/0.15.0...0.15.1
 [diff: 0.14.0-0.15.0]: https://github.com/Luzkan/CryptoImage/compare/0.14.0...0.15.0
 [diff: 0.13.0-0.14.0]: https://github.com/Luzkan/CryptoImage/compare/0.13.0...0.14.0
@@ -420,6 +434,7 @@ Added:
 [diff: 0.1.0-0.1.1]: https://github.com/Luzkan/CryptoImage/compare/0.1.0...0.1.1
 [diff: 0.0.1-0.1.0]: https://github.com/Luzkan/CryptoImage/compare/0.0.1...0.1.0
 [diff: 0.0.0-0.0.1]: https://github.com/Luzkan/CryptoImage/compare/0.0.0...0.0.1
+[0.15.2]: https://github.com/Luzkan/CryptoImage/releases/tag/0.15.2
 [0.15.1]: https://github.com/Luzkan/CryptoImage/releases/tag/0.15.1
 [0.15.0]: https://github.com/Luzkan/CryptoImage/releases/tag/0.15.0
 [0.14.0]: https://github.com/Luzkan/CryptoImage/releases/tag/0.14.0